requirements.txt
python-dotenv
requests
PyYAML
colorama